LayoutTests/imported/w3c/web-platform-tests/uievents/textInput/enter-textarea-contenteditable-expected.txt

Press Enter in each text field below.





FAIL textInput: Enter key for textarea and contenteditable, <textarea> assert_equals: expected 1 but got 0
FAIL textInput: Enter key for textarea and contenteditable, <div contenteditable="true"> assert_equals: expected 1 but got 0
